Add files via upload
These are the rough implementations of the login functionality of our database. It is working. I just need another half another to make it fully functional. Right now it will properly conclude whether there are users in the table if BOTH the username and password are entered correctly. However if someone makes a mistake and one of the details is misspelled it will just say the user doesn't exists and tell you to create one. I know how to fix it and will work on it when we meet up.